+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for `module_tags`
+- Implement support for resource servers through the
+  `aws_cognito_resource_server` resource
 
 ## [0.9.0]